🗺️ Atlas: Sub-modularize Core API Blobs
